+++ backend/api/tags/serializers.py
@@ -1,0 +1,18 @@
+from rest_framework import serializers
+
+from tags.models import Tag
+
+
+class TagSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Tag
+        fields = ["id", "name", "description", "question_count"]
+
+
+class TagStringSerializer(serializers.ModelSerializer):
+    def to_representation(self, value):
+        return value.name
+
+    class Meta:
+        model = Tag
+        fields = ["name"]
