+++ backend/api/views.py
@@ -13,14 +13,13 @@ from rest_framework.response import Response
 
 from api import constants, utilities_sendinblue
 from api.filters import QuestionFilter, QuizFilter
-from api.models import Question, Quiz, Tag
+from api.models import Question, Quiz
 from api.serializers import (
     QuestionDifficultyChoiceSerializer,
     QuestionFullStringSerializer,
     QuestionSerializer,
     QuizSerializer,
     SimpleChoiceSerializer,
-    TagSerializer,
 )
 
 
@@ -132,15 +131,6 @@ def question_random(request):
         serializer = QuestionSerializer(question_random)
 
     return Response(serializer.data)
-
-
-class TagViewSet(mixins.ListModelMixin, viewsets.GenericViewSet):
-    queryset = Tag.objects.all().order_by("name")
-    serializer_class = TagSerializer
-
-    @extend_schema(summary="Lister tous les tags", tags=[Tag._meta.verbose_name_plural])
-    def list(self, request, *args, **kwargs):
-        return super().list(request, args, kwargs)
 
 
 @api_view(["GET"])
